+++ packages/ra-core/src/core/useConfigureAdminRouterFromChildren.spec.tsx
@@ -7,6 +7,7 @@ import { CoreAdminContext } from './CoreAdminContext';
 import { CoreAdminRoutes } from './CoreAdminRoutes';
 import { Resource } from './Resource';
 import { CoreLayoutProps } from '../types';
+import { AuthProvider, ResourceProps } from '..';
 
 const ResourceDefinitionsTestComponent = () => {
     const definitions = useResourceDefinitions();
@@ -14,7 +15,9 @@ const ResourceDefinitionsTestComponent = () => {
     return (
         <ul>
             {Object.keys(definitions).map(key => (
-                <li key={key}>{JSON.stringify(definitions[key])}</li>
+                <li key={key} data-resource={key}>
+                    {JSON.stringify(definitions[key])}
+                </li>
             ))}
         </ul>
     );
@@ -53,14 +56,108 @@ const TestedComponent = ({ role }) => {
     );
 };
 
+const ResourceWithPermissions = (props: ResourceProps) => (
+    <Resource {...props} />
+);
+ResourceWithPermissions.raName = 'Resource';
+ResourceWithPermissions.registerResource = (
+    { create, edit, icon, list, name, options, show }: ResourceProps,
+    permissions: any
+) => ({
+    name,
+    options,
+    hasList: !!list && permissions && permissions[name]?.list,
+    hasCreate: !!create && permissions && permissions[name]?.create,
+    hasEdit: !!edit && permissions && permissions[name]?.edit,
+    hasShow: !!show && permissions && permissions[name]?.show,
+    icon,
+});
+
+const TestedComponentWithPermissions = () => {
+    const history = createMemoryHistory();
+    const authProvider: AuthProvider = {
+        login: () => Promise.resolve(),
+        logout: () => Promise.resolve(),
+        checkAuth: () => Promise.resolve(),
+        checkError: () => Promise.resolve(),
+        getPermissions: () =>
+            Promise.resolve({
+                posts: {
+                    list: true,
+                    create: true,
+                    edit: true,
+                    show: true,
+                },
+                comments: {
+                    list: true,
+                    create: false,
+                    edit: false,
+                    show: true,
+                },
+                users: {
+                    list: true,
+                    create: false,
+                    edit: false,
+                    show: false,
+                },
+            }),
+    };
+
+    return (
+        <CoreAdminContext authProvider={authProvider} history={history}>
+            <CoreAdminRoutes
+                layout={MyLayout}
+                catchAll={CatchAll}
+                loading={Loading}
+            >
+                <ResourceWithPermissions
+                    name="posts"
+                    list={<div />}
+                    create={<div />}
+                    edit={<div />}
+                    show={<div />}
+                />
+                <ResourceWithPermissions
+                    name="comments"
+                    list={<div />}
+                    create={<div />}
+                    edit={<div />}
+                    show={<div />}
+                />
+                <ResourceWithPermissions
+                    name="users"
+                    list={<div />}
+                    create={<div />}
+                    edit={<div />}
+                    show={<div />}
+                />
+            </CoreAdminRoutes>
+        </CoreAdminContext>
+    );
+};
+
 const expectResource = (resource: string) =>
     expect(screen.queryByText(`"name":"${resource}"`, { exact: false }));
+
+const expectResourceView = (
+    resource: string,
+    view: 'list' | 'create' | 'edit' | 'show'
+) =>
+    expect(
+        screen.queryByText(
+            `"has${view.at(0).toUpperCase()}${view.substring(1)}":true`,
+            {
+                selector: `[data-resource=${resource}]`,
+                exact: false,
+            }
+        )
+    );
 
 describe('useConfigureAdminRouterFromChildren', () => {
     it('should always load static resources', async () => {
         render(<TestedComponent role="guest" />);
         await waitFor(() => expect(screen.queryByText('Loading')).toBeNull());
-        expectResource('posts').not.toBeNull();
+        await waitFor(() => expectResource('posts').not.toBeNull());
         expectResource('comments').not.toBeNull();
         expectResource('user').toBeNull();
         expectResource('admin').toBeNull();
@@ -70,6 +167,23 @@ describe('useConfigureAdminRouterFromChildren', () => {
         await waitFor(() => expect(screen.queryByText('Loading')).toBeNull());
         expectResource('user').not.toBeNull();
         expectResource('admin').not.toBeNull();
+    });
+    it('should call registerResource with the permissions', async () => {
+        render(<TestedComponentWithPermissions />);
+        await waitFor(() => expect(screen.queryByText('Loading')).toBeNull());
+
+        expectResourceView('posts', 'list').not.toBeNull();
+        expectResourceView('posts', 'create').not.toBeNull();
+        expectResourceView('posts', 'edit').not.toBeNull();
+        expectResourceView('posts', 'show').not.toBeNull();
+        expectResourceView('comments', 'list').not.toBeNull();
+        expectResourceView('comments', 'create').toBeNull();
+        expectResourceView('comments', 'edit').toBeNull();
+        expectResourceView('comments', 'show').not.toBeNull();
+        expectResourceView('users', 'list').not.toBeNull();
+        expectResourceView('users', 'create').toBeNull();
+        expectResourceView('users', 'edit').toBeNull();
+        expectResourceView('users', 'show').toBeNull();
     });
     it('should allow adding new resource after the first render', async () => {
         const { rerender } = render(<TestedComponent role="user" />);