+++ packages/ra-core/src/util/getFieldLabelTranslationArgs.spec.ts
@@ -43,11 +43,22 @@ describe('getFieldLabelTranslationArgs', () => {
         ]);
     });
 
-    it('should return the source and resource as translate key', () =>
+    it('should return the source and resource as translate key', () => {
         expect(
             getFieldLabelTranslationArgs({
                 resource: 'posts',
                 source: 'title',
             })
-        ).toEqual([`resources.posts.fields.title`, { _: 'Title' }]));
+        ).toEqual([`resources.posts.fields.title`, { _: 'Title' }]);
+    });
+
+    it('should accept use the parentSource to build the translation key if provided', () => {
+        expect(
+            getFieldLabelTranslationArgs({
+                resource: 'posts',
+                source: 'url',
+                parentSource: 'backlinks',
+            })
+        ).toEqual([`resources.posts.fields.backlinks.url`, { _: 'Url' }]);
+    });
 });