+++ packages/ra-input-rich-text/src/RichTextInputToolbar.tsx
@@ -34,7 +34,7 @@ import {
  *     <RichTextInput
  *         toolbar={(
  *             <RichTextInputToolbar>
- *                 <RichTextInputLevelSelect size={size} />
+ *                 <LevelSelect size={size} />
  *                 <FormatButtons size={size} />
  *                 <ListButtons size={size} />
  *                 <LinkButtons size={size} />