+++ packages/ra-no-code/src/ui/index.ts
@@ -2,5 +2,5 @@ export * from './Layout';
 export * from './Menu';
 export * from './UserMenu';
 export * from './ExitApplicationMenu';
-export * from './Ready';
 export * from './ImportResourceDialog';
+export * from './Ready';