+++ packages/ra-ui-materialui/src/input/ArrayInput/SimpleFormIterator.spec.tsx
@@ -48,7 +48,7 @@ describe('<SimpleFormIterator />', () => {
             </Wrapper>
         );
         const inputElements = screen.queryAllByLabelText(
-            'resources.undefined.fields.email'
+            'resources.undefined.fields.emails.email'
         );
         expect(inputElements).toHaveLength(2);
         expect((inputElements[0] as HTMLInputElement).disabled).toBeFalsy();
@@ -75,7 +75,7 @@ describe('<SimpleFormIterator />', () => {
             </Wrapper>
         );
         const inputElements = screen.queryAllByLabelText(
-            'resources.undefined.fields.email'
+            'resources.undefined.fields.emails.email'
         );
         expect(inputElements).toHaveLength(2);
         expect((inputElements[0] as HTMLInputElement).disabled).toBeTruthy();
@@ -102,7 +102,7 @@ describe('<SimpleFormIterator />', () => {
             </Wrapper>
         );
         const inputElements = screen.queryAllByLabelText(
-            'resources.undefined.fields.email'
+            'resources.undefined.fields.emails.email'
         );
         expect(inputElements).toHaveLength(2);
         expect((inputElements[0] as HTMLInputElement).disabled).toBeTruthy();
@@ -180,6 +180,31 @@ describe('<SimpleFormIterator />', () => {
         expect(screen.queryAllByText('ra.action.remove').length).toBe(0);
     });
 
+    it('should not display remove button if disableRemove return value is truthy', () => {
+        render(
+            <Wrapper>
+                <SimpleForm
+                    record={{
+                        id: 'whatever',
+                        emails: [{ email: 'badEmail' }, { email: '' }],
+                    }}
+                >
+                    <ArrayInput source="emails">
+                        <SimpleFormIterator
+                            disableRemove={record => {
+                                return record.email === 'badEmail';
+                            }}
+                        >
+                            <TextInput source="email" />
+                        </SimpleFormIterator>
+                    </ArrayInput>
+                </SimpleForm>
+            </Wrapper>
+        );
+
+        expect(screen.queryAllByText('ra.action.remove').length).toBe(1);
+    });
+
     it('should not display remove button if disabled is truthy', () => {
         render(
             <Wrapper>
@@ -221,7 +246,7 @@ describe('<SimpleFormIterator />', () => {
         fireEvent.click(addItemElement);
         await waitFor(() => {
             const inputElements = screen.queryAllByLabelText(
-                'resources.undefined.fields.email'
+                'resources.undefined.fields.emails.email'
             );
 
             expect(inputElements.length).toBe(1);
@@ -230,14 +255,14 @@ describe('<SimpleFormIterator />', () => {
         fireEvent.click(addItemElement);
         await waitFor(() => {
             const inputElements = screen.queryAllByLabelText(
-                'resources.undefined.fields.email'
+                'resources.undefined.fields.emails.email'
             );
 
             expect(inputElements.length).toBe(2);
         });
 
         const inputElements = screen.queryAllByLabelText(
-            'resources.undefined.fields.email'
+            'resources.undefined.fields.emails.email'
         ) as HTMLInputElement[];
 
         expect(
@@ -339,7 +364,7 @@ describe('<SimpleFormIterator />', () => {
         );
 
         const inputElements = screen.queryAllByLabelText(
-            'resources.undefined.fields.email'
+            'resources.undefined.fields.emails.email'
         ) as HTMLInputElement[];
 
         expect(
@@ -356,7 +381,7 @@ describe('<SimpleFormIterator />', () => {
         fireEvent.click(removeFirstButton);
         await waitFor(() => {
             const inputElements = screen.queryAllByLabelText(
-                'resources.undefined.fields.email'
+                'resources.undefined.fields.emails.email'
             ) as HTMLInputElement[];
 
             expect(
@@ -383,7 +408,7 @@ describe('<SimpleFormIterator />', () => {
         );
 
         const inputElements = screen.queryAllByLabelText(
-            'resources.undefined.fields.email'
+            'resources.undefined.fields.emails.email'
         ) as HTMLInputElement[];
 
         expect(
@@ -399,7 +424,7 @@ describe('<SimpleFormIterator />', () => {
         fireEvent.click(moveDownFirstButton[0]);
         await waitFor(() => {
             const inputElements = screen.queryAllByLabelText(
-                'resources.undefined.fields.email'
+                'resources.undefined.fields.emails.email'
             ) as HTMLInputElement[];
 
             expect(
@@ -414,7 +439,7 @@ describe('<SimpleFormIterator />', () => {
         fireEvent.click(moveUpButton[1]);
         await waitFor(() => {
             const inputElements = screen.queryAllByLabelText(
-                'resources.undefined.fields.email'
+                'resources.undefined.fields.emails.email'
             ) as HTMLInputElement[];
 
             expect(