+++ packages/ra-ui-materialui/src/input/ArrayInput/SimpleFormIterator.tsx
@@ -12,11 +12,14 @@ import { styled } from '@mui/material';
 import clsx from 'clsx';
 import get from 'lodash/get';
 import PropTypes from 'prop-types';
-import { RaRecord } from 'ra-core';
+import { RaRecord, useRecordContext } from 'ra-core';
 import { UseFieldArrayReturn } from 'react-hook-form';
 
 import { useArrayInput } from './useArrayInput';
-import { SimpleFormIteratorClasses } from './useSimpleFormIteratorStyles';
+import {
+    SimpleFormIteratorClasses,
+    SimpleFormIteratorPrefix,
+} from './useSimpleFormIteratorStyles';
 import { SimpleFormIteratorContext } from './SimpleFormIteratorContext';
 import {
     DisableRemoveFunction,
@@ -33,7 +36,6 @@ export const SimpleFormIterator = (props: SimpleFormIteratorProps) => {
         reOrderButtons = <DefaultReOrderButtons />,
         children,
         className,
-        record,
         resource,
         source,
         disabled,
@@ -43,6 +45,7 @@ export const SimpleFormIterator = (props: SimpleFormIteratorProps) => {
         getItemLabel = DefaultLabelFn,
     } = props;
     const { append, fields, move, remove } = useArrayInput(props);
+    const record = useRecordContext(props);
 
     const removeField = useCallback(
         (index: number) => {
@@ -83,8 +86,9 @@ export const SimpleFormIterator = (props: SimpleFormIteratorProps) => {
             add: addField,
             remove: removeField,
             reOrder: handleReorder,
+            source,
         }),
-        [fields.length, addField, removeField, handleReorder]
+        [addField, fields.length, handleReorder, removeField, source]
     );
     return fields ? (
         <SimpleFormIteratorContext.Provider value={context}>
@@ -176,7 +180,10 @@ export interface SimpleFormIteratorProps extends Partial<UseFieldArrayReturn> {
     source?: string;
 }
 
-const Root = styled('ul')(({ theme }) => ({
+const Root = styled('ul', {
+    name: SimpleFormIteratorPrefix,
+    overridesResolver: (props, styles) => styles.root,
+})(({ theme }) => ({
     padding: 0,
     marginBottom: 0,
     '& > li:last-child': {