+++ packages/ra-ui-materialui/src/input/NumberInput.tsx
@@ -11,6 +11,9 @@ import { sanitizeInputRestProps } from './sanitizeInputRestProps';
 /**
  * An Input component for a number
  *
+ * Due to limitations in React controlled components and number formatting,
+ * this input only updates the form value on blur.
+ *
  * @example
  * <NumberInput source="nb_views" />
  *
@@ -18,25 +21,24 @@ import { sanitizeInputRestProps } from './sanitizeInputRestProps';
  * @example
  * <NumberInput source="nb_views" step={1} />
  *
- * The object passed as `options` props is passed to the MUI <TextField> component
  */
 export const NumberInput = ({
     className,
-    defaultValue = '',
+    defaultValue = null,
     format = convertNumberToString,
     helperText,
     label,
-    margin = 'dense',
+    margin,
     onBlur,
     onChange,
-    parse = convertStringToNumber,
+    parse,
     resource,
     source,
     step,
     min,
     max,
     validate,
-    variant = 'filled',
+    variant,
     inputProps: overrideInputProps,
     ...rest
 }: NumberInputProps) => {
@@ -48,10 +50,6 @@ export const NumberInput = ({
         isRequired,
     } = useInput({
         defaultValue,
-        format,
-        onBlur,
-        onChange,
-        parse,
         resource,
         source,
         validate,
@@ -60,10 +58,64 @@ export const NumberInput = ({
 
     const inputProps = { ...overrideInputProps, step, min, max };
 
+    // This is a controlled input that doesn't transform the user input on change.
+    // The user input is only turned into a number on blur.
+    // This is to allow transitory values like '1.0' that will lead to '1.02'
+
+    // text typed by the user and displayed in the input, unparsed
+    const [value, setValue] = React.useState(format(field.value));
+
+    // update the input text when the record changes
+    React.useEffect(() => {
+        const stringValue = format(field.value);
+        setValue(value => (value !== stringValue ? stringValue : value));
+    }, [field.value]); // eslint-disable-line react-hooks/exhaustive-deps
+
+    // update the input text when the user types in the input
+    const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        if (onChange) {
+            onChange(event);
+        }
+        if (
+            typeof event.target === 'undefined' ||
+            typeof event.target.value === 'undefined'
+        ) {
+            return;
+        }
+        setValue(event.target.value);
+    };
+
+    // set the numeric value on the form on blur
+    const handleBlur = (...event: any[]) => {
+        if (onBlur) {
+            onBlur(...event);
+        }
+        const eventParam = event[0] as React.FocusEvent<HTMLInputElement>;
+        if (
+            typeof eventParam.target === 'undefined' ||
+            typeof eventParam.target.value === 'undefined'
+        ) {
+            return;
+        }
+        const target = eventParam.target;
+        const newValue = target.valueAsNumber
+            ? parse
+                ? parse(target.valueAsNumber)
+                : target.valueAsNumber
+            : parse
+            ? parse(target.value)
+            : convertStringToNumber(target.value);
+        field.onChange(newValue);
+    };
+
     return (
         <TextField
             id={id}
             {...field}
+            // override the react-hook-form value, onChange and onBlur props
+            value={value}
+            onChange={handleChange}
+            onBlur={handleBlur}
             className={clsx('ra-input', `ra-input-${source}`, className)}
             type="number"
             size="small"
@@ -117,11 +169,13 @@ export interface NumberInputProps
 }
 
 const convertStringToNumber = value => {
+    if (value == null || value === '') {
+        return null;
+    }
     const float = parseFloat(value);
 
-    return isNaN(float) ? null : float;
+    return isNaN(float) ? 0 : float;
 };
 
-const convertNumberToString = value => {
-    return value === null ? '' : value.toString();
-};
+const convertNumberToString = value =>
+    value == null || isNaN(value) ? '' : value.toString();